requirements.txt
Flask>=2.2
python-dotenv>=0.21
qrcode[pil]
